feat(reporter): add browserName function and use where required
